refactor: update user service dependency injection to use async Redis and streamline service retrieval
